ExPool v0.0.1 ExPool.Pool

Pool

Functions

check_in(pool, worker)

Specs

check_in(pool :: pid, worker :: pid) :: :ok

Check-in a worker into the pool.

check_out(pool)

Specs

check_out(pool :: pid) :: worker :: pid

Check-out a worker from the pool. It blocks until one is available.

start_link(opts \\ [])

Specs

start_link(opts :: [Keyword]) :: Supervisor.on_start

Starts a new pool GenServer.

Options:

:name - (Optional) The name of the pool

The rest of the options will be passed to the internal state as pool configuration (for more information about the available options check ExPool.Pool.State.new/1).