ExPool v0.0.1 ExPool.Pool.State

The internal state of a pool.

Types

t :: %ExPool.Pool.State{worker_mod: atom, size: pos_integer, sup: pid, workers: [pid], queue: any}

Functions

enqueue(state, from)

Specs

enqueue(State.t, from :: any) :: State.t

Adds a pair (pid, ref) to the queue.

get_worker(state)

Specs

get_worker(State.t) ::
  {:ok, {pid, State.t}} |
  {:empty, State.t}

Get a worker and remove it from the workers list.

new(config)

Specs

new(config :: [Keyword]) :: State.t

Creates a new pool state with the given configuration.

Configuration options

:worker_mod - The worker module. It has to fit in a supervision tree (like a GenServer).

:size - The size of the pool (default 5).

pop_from_queue(state)

Specs

pop_from_queue(State.t) ::
  {:ok, {item :: any, State.t}} |
  {:empty, State.t}

Pops a pair (pid, ref) from the queue.

put_worker(state, worker)

Specs

put_worker(State.t, pid) :: State.t

Add a worker to the workers list.